requirements.txt
# FastAPI and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# RAG and AI Related
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
google-generativeai>=0.3.0

# Vector Database
chromadb>=0.4.0

# Embeddings and Text Processing
sentence-transformers>=2.2.0

# Data Processing
numpy>=1.24.0

# Environment Variables and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
requests>=2.31.0
aiofiles>=23.0.0

# Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0